Setup general yap lib
