+++ src/app/harbor/shop/shopkeeper.js
@@ -1,7 +1,12 @@
 import { useState } from 'react'
 
 import { Howl } from 'howler'
-import { sample, shopGreeting, shopNoMoney } from '../../../../lib/flavor'
+import { sample, shopGreeting, shopNoMoney, shopGetOut } from '../../../../lib/flavor'
+import { yap } from '../../../../lib/yap'
+
+async function shopkeeperSay(text) {
+  await yap(text)
+}
 
 const bellSoundUrls = [
   'https://cloud-dx9y4rk8f-hack-club-bot.vercel.app/0ding-2-90199_audio.mp4',
@@ -12,20 +17,23 @@ const bellSoundUrls = [
 ]
 const bellSounds = bellSoundUrls.map((url) => new Howl({ src: [url] }))
 
-export const ShopkeeperComponent = ({ balance: number }) => {
+export const ShopkeeperComponent = ({ balance: number, cursed: boolean }) => {
   const [atCounter, setAtCounter] = useState(false)
   const [bellIndex, setBellIndex] = useState(0)
   const [bellClickCount, setBellClickCount] = useState(0)
   const [continuousBellClicks, setContinuousBellClicks] = useState(0)
 
-  const handleServiceBellClick = () => {
-    bellSounds[bellIndex].play()
+  const handleServiceBellClick = async () => {
     setBellIndex(Math.floor(Math.random() * bellSounds.length))
     setBellClickCount(bellClickCount + 1)
     setContinuousBellClicks(continuousBellClicks + 1)
 
+    bellSounds[bellIndex].play()
+
+    await new Promise((resolve) => setTimeout(resolve, 1000))
+
     if (balance == 0) {
-      shopkeeperSay(
+      await shopkeeperSay(
         sample(shopGreeting) +
           ' ' +
           sample(shopNoMoney) +
